COLORADO-1304 Add test for playable derived types when base type has no viewer
There is a difference between what the viewer considers a viewable type and the
ability to view a type.  There are many more viewable types than there are viewers
to view them.  In the case of a viewable type without a viewer, there is an
assumption of the existence of a derived type in the object that can be viewed.

The viewer previously used the DSID of the viewable type without knowing if that
type had a viewer.  The fix tests the viewable type to see if there is a viewer
for it and if there is not then the list of derived types is tested to see if
there is a viewer for one of the derived types.  If such a viewer exists for a
derived type then the DSID of the derived type is used.