COLORADO-1559 Added support for datastreams to respect XACML policies
The desired behaviour for viewing datastreams using XACML policies is to
consider a datastream viewable if either it or any of its derived types
has permission to be viewed by the user.

The other restrictions on viewability are not affected (do not view
thumbnails, etc.)

The second part of the fix was to retrieve a viewable datastream for a
given DSID.  The test in this case is to test if the given DSID both
has a viewer and the user has permission to view the DSID.  If this test
fails then the derived types are all tested for permission and an
existing viewer.

The final change to make this work was to exclude the thumbnail relation
(hasThumbnail) fro the list of derived datastreams.