COLORADO-1203 Change datastream Add button behaviour to permit multi-file ingest
This update sends the "Add" button click to a drupal path that will contain a
form capable of ingesting multiple files into datastreams of the current object.

The drupal path is set to <host>/islandora_co_manager/ingest_files/<pid> where
<pid> is the pid of the current object or, in the case of the root collection,
the pid will be set to the string "undefined".