COLORADO-1672 Send a viewable DSID to the viewer and use that for viewing
The viewer code assumes that the DSID passed to it will be viewable and this
is not always the case.  The fix for this problem is to pass a viewable DSID
with each object that is sent to the viewer.  The viewer uses this DSID when
it gets a datastream to display.

Minor changes to datastructures that are passed to the viewer from the Drupal
module as well as associated minor changes to use the correct DSID inside
of the viewer code.