COLORADO-681 Changed CSS for metadata titles in object view
The CSS for the titles in the object view is generated inline in
AjaxHandler.inc.  These were changed to be more in line with what was
desired by the client.

An additional minor change was made to the footer of the view to move
it to its own table.  This allowed the metadata titles to remain a
constant size regardless of the size of the rest of the data
in the table.  If the footer was significantly larger than the combined
length of the longest title and longest value fields then the title
column would be stretched an inordinate amount.