COLORADO-942 Various JavaScript errors were causing IE7 to fail to work
There were a number of JavaScript errors in the code we create for the viewer.
The errors that caused IE7 to fail involved trailing commas after the last item
in a list.  One of the files that had these errors was generated by Ext Designer
and, as such, had to be edited by hand to remove the errors.  It is VERY likely
that the errors will reoccur if the program is used to alter the JavaScript.  A
lare warning has been put in the file that contained the generated errors.

The remainder of the fixed JavaScript tended to be missing semicolons at the end
of statements.  There was also n error in one of the PHP files that sent
JavaScript directly to the browser.