COLORADO-1597 Lengthen size of display element to make pages pane visible
There apepars to be an overarching problem with the way that the height of
files pane is being generated in Forefox.  The most expidicious fix is to
ensure that the minimum height of the entire overview tab is large enough
to contain a full complement of 5 files.  I have done this and made the
minimum height larger.