COLORADO-1615 Fixed problem with redirect datastreams not being displayed
The viewer currently returns the data in a datastream (get_datastream_dissemination)
and when a datastream is a redirect stream the data is a URL pointing to the
actual data. The fix for this is to check the content type field of the datastream
and if it is a redirect field then download the actual data with a curl call.