COLORADO-1615 Fixed my stupidity and added a HTML viewer
I left some debugging code in the XML viewer that effectively caused the
XML processing to be skipped.  Very stupid.  I fixed that error and added
in a viewer that allowed HTML files to be viewed in a manner that was very
similar to regular a regular web browser.  This was in response to the
fact that the initial cause of the XML bug came from a redirect object
that was, in fact, a HTML object.  Even after fixing the redirect problem
the HTML file would not display correctly becasue it was MIME-typed as
text/xml but was not well-formed and could not be parsed.