COLORADO-1911 Prefer display of derived DSIDs over original DSID
The current behaviour in the viewer is to test the original datastream to see
if it can be displayed.  If it can then it is used as the viewable DSID.  The
desired behaviour is to first test to see if there is a derived DSID that is
displayable and use that DSID as it is likely to have been created with the
primary goal of being in a format that is better able to be displayed.  That
behaviour is now the default for the browser.